:recycle: Move local_cliffords stuff into ZXGraphState
